requirements.txt
streamlit
openai
googlemaps
requests
python-dotenv
pandas
matplotlib
seaborn
